Fix to flexslider so previous image isn't shown

+++ assets/js/flexslider/jquery.flexslider.js
@@ -916,9 +916,9 @@
 
       if (slider.transitions) {
         if (slider.isFirefox) {
-          target = (vertical) ? "translate3d(0," + target + ",0)" : "translate3d(" + (parseInt(target)+'px') + ",0,0)";
+          target = (vertical) ? "translate3d(0," + target + ",0)" : "translate3d(" + (parseFloat(target)+'px') + ",0,0)";
         } else {
-          target = (vertical) ? "translate3d(0," + target + ",0)" : "translate3d(" + ((slider.vars.rtl?-1:1)*parseInt(target)+'px') + ",0,0)";
+          target = (vertical) ? "translate3d(0," + target + ",0)" : "translate3d(" + ((slider.vars.rtl?-1:1)*parseFloat(target)+'px') + ",0,0)";
         }
         dur = (dur !== undefined) ? (dur/1000) + "s" : "0s";
         slider.container.css("-" + slider.pfx + "-transition-duration", dur);
